(AddVersioning)
- ensure to add versioning to the application
[Finishes #166467430]
