+++ spec/controllers/v1/questions_controller_spec.rb
@@ -23,7 +23,7 @@ require 'rails_helper'
 # removed from Rails core in Rails 5, but can be added back in via the
 # `rails-controller-testing` gem.
 
-RSpec.describe QuestionsController, type: :controller do
+RSpec.describe V1::QuestionsController, type: :controller do
 
   # This should return the minimal set of attributes required to create a valid
   # Question. As you add validations to Question, be sure to
